requirements.txt
flask~=2.0.2
flask-sqlalchemy
httpx~=0.20.0

pip~=21.3
toml~=0.10.2
wheel~=0.36.2
cryptography~=3.4.8
setuptools~=57.0.0
python-dotenv~=0.19.1